LayoutTests/platform/ios-simulator/fast/css/parsing-css-not-1-expected.txt

Test the parsing of :not(selectorList) for querySelector and style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


1 valid selectors
PASS document.querySelector(":not(*)") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(*)"
PASS document.querySelector(":not(*    )") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(*)"
PASS document.querySelector(":not(foobar)") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(foobar)"
PASS document.querySelector(":not(foobar    )") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(foobar)"
PASS document.querySelector(":not(#id)") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(#id)"
PASS document.querySelector(":not(#id    )") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(#id)"
PASS document.querySelector(":not(.class)") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(.class)"
PASS document.querySelector(":not(.class    )") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(.class)"
PASS document.querySelector(":not(:first-child)") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(:first-child)"
PASS document.querySelector(":not(:first-child    )") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(:first-child)"
PASS document.querySelector(":not(:last-child)") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(:last-child)"
PASS document.querySelector(":not(:last-child    )") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(:last-child)"
PASS document.querySelector(":not(:visited)") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(:visited)"
PASS document.querySelector(":not(:visited    )") did not throw exception.
PASS document.getElementById('style-container').sheet.cssRules.length is 1
PASS document.getElementById('style-container').sheet.cssRules[0].selectorText is ":not(:visited)"
FAIL document.querySelector(":not(foobar#id.class)") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(foobar#id.class). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(foobar#id.class    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(foobar#id.class). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(.class:not(.notclass))") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(.class:not(.notclass)). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(.class:not(.notclass)    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(.class:not(.notclass)). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(.class:matches(.notclass))") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(.class:matches(.notclass)). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(.class:matches(.notclass)    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(.class:matches(.notclass)). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(#id:empty)") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(#id:empty). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(#id:empty    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(#id:empty). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(a > b)") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(a > b). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(a > b    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(a > b). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(a b)") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(a b). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(a b    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(a b). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(a + b)") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(a + b). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(a + b    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(a + b). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(a ~ b)") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(a ~ b). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(a ~ b    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(a ~ b). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(a + b > c ~ d e + g)") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(a + b > c ~ d e + g). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(a + b > c ~ d e + g    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(a + b > c ~ d e + g). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(:-webkit-any(.selector,#tama,#hanayo,#midoriko))") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(:-webkit-any(.selector,#tama,#hanayo,#midoriko)). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(:-webkit-any(.selector,#tama,#hanayo,#midoriko)    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(:-webkit-any(.selector,#tama,#hanayo,#midoriko)). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(:not(:link))") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(:not(:link)). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(:not(:link)    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(:not(:link)). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(:nth-child(even of a, b, c))") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(:nth-child(even of a, b, c)). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.querySelector(":not(:nth-child(even of a, b, c)    )") should not throw exception. Threw exception Error: SyntaxError: DOM Exception 12.
FAIL document.getElementById('style-container').sheet.cssRules.length should be 1. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :not(:nth-child(even of a, b, c)). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
PASS successfullyParsed is true

TEST COMPLETE
